Ticket(s): #5259 - Ability to toggle on/off the list of tools in the right-click context menu - Added configuration page to hide all tools from the context menu. - Added context menu filtering to the ImageViewerComponent.  Doing this sort of thing on a more global scale proved difficult with the current actions architecture, so I gave up for now. - Put tools that will be excluded from the context menu on the global toolbar (and menus in some cases). - Added action filter rules (currently hidden) as a setting; currently only used for the context menu. - Moved around some of the actions on the global toolbar. - By default, the functionality is disabled unless a valid setting is applied. - Also did some refactoring that should make it easier to customize the viewer (for MPR, Fusion, as an example) without having to inherit from ImageViewerComponent.  Added the IViewerSetupHelper that can be passed to the viewer's constructor, allowing things like the LayoutManager and ToolSet to be provided *to* the viewer. - Fixed a small problem in drop-down toolbar controls where the icons were being scaled down.
Schema/Configuration Changes: YES
- New settings class ImageViewer.Configuration.ToolSettings contains action filter rules.
- By default, the setting is empty, and we will have to enable it for JDMI specifically.

+++ ImageViewer/Tools/Standard/DeleteAnnotationsTool.cs
@@ -1,6 +1,6 @@
 ﻿#region License
 
-// Copyright (c) 2009, ClearCanvas Inc.
+// Copyright (c) 2010, ClearCanvas Inc.
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification, 
@@ -129,6 +129,13 @@ namespace ClearCanvas.ImageViewer.Tools.Standard
 	[VisibleStateObserver("deleteall", "DeleteAllVisible", "DeleteAllVisibleChanged")]
 	[IconSet("deleteall", IconScheme.Colour, "DeleteAllAnnotationsToolSmall.png", "DeleteAllAnnotationsToolMedium.png", "DeleteAllAnnotationsToolLarge.png")]
 	[GroupHint("deleteall", "Tools.Image.Annotations.DeleteAll")]
+
+	[ButtonAction("deleteallToolbar", "global-toolbars/ToolbarAnnotation/ToolbarDeleteAllAnnotations", "DeleteAll")]
+	[EnabledStateObserver("deleteallToolbar", "DeleteAllVisible", "DeleteAllVisibleChanged")]
+	[IconSet("deleteallToolbar", IconScheme.Colour, "DeleteAllAnnotationsToolSmall.png", "DeleteAllAnnotationsToolMedium.png", "DeleteAllAnnotationsToolLarge.png")]
+	[Tooltip("deleteallToolbar", "TooltipDeleteAllAnnotations")]
+	[GroupHint("deleteallToolbar", "Tools.Image.Annotations.DeleteAll")]
+
 	[ExtensionOf(typeof(ImageViewerToolExtensionPoint))]
 	public class DeleteAllAnnotationsTool : ImageViewerTool
 	{
@@ -156,7 +163,7 @@ namespace ClearCanvas.ImageViewer.Tools.Standard
 
 		public void DeleteAll()
 		{
-			if (base.SelectedPresentationImage != null)
+			if (DeleteAllVisible && base.SelectedPresentationImage != null)
 				DeleteAll(this.SelectedPresentationImage);
 		}
